Tipos de tarjeta, verificacion de edad, arreglo tarjetas

+++ src/main/java/Modelo/ArregloCliente.java
@@ -41,7 +41,7 @@ public class ArregloCliente {
         return true;
     }
 
-    public boolean registrarTarjeta(int numero, String nombre, String fecha, int cvv, double saldo) {
+    public boolean registrarTarjeta(long numero, String nombre, String fecha, int cvv, double saldo) {
         if (clienteActual == null) {
             return false;
         }
